// web-client/src/types.d.ts
export interface UserType {
  id: number;
  name: string;
  email: string;
  password: string;
  registrationDate: string;
  country: string | null;
  profileImage: string | null;
  active: boolean;
}